requirements.txt
appdirs>=1.4.4
beautifulsoup4>=4.11.1
lxml>=4.9.1
yfinance>=0.2.38
flask
pandas
google-generativeai
langchain
openai
faiss-cpu 